BB-190: Rename all react component files to kebab case
