+++ src/client/components/entity-link.jsx
