+++ src/client/components/forms/parts/publication-data.jsx
